1.3.0
New feature: It is now possible to move when an unpaused menu is open, if using a keyboard (disabled by default, can be enabled in the .ini)
It is now possible to slow down the time while an unpaused menu is open. (disabled by default, can be enabled in the .ini)
Fixed some issues with ServeTime and SaveGame console commands.
Some internal changes & code cleanup.