1.4.0
Menus will no longer auto close if the player is interacting with a disabled container, as long as it does not get disabled while it's open.
Controllers are now supported when the movement is enabled in the menus.
Fixed controls sometimes not working properly in favorites menu and book menu.
Added .ini setting bEnableGamepadCameraMove. If enabled, it is possible to rotate the camera when using a controller and a menu is open.
Added .ini setting bEnableCursorCameraMove, if enabled, it is possible to move the camera by moving the cursor to the edge of the screen (similiar to dialogue menu). Its speed is configurable.
The cursor will now get centered by default every time when opening a menu.
Fixed random CTDs when remapping controls by making the game to be in a paused state while remapping.
Some optimizations.